Update route for bids
Made comments, for the purpose of clarification.
Updated the add function for the new schema.
Changed the search function, now we can search for a bid, by its user's id.
Like this, we can attach bids for users.
Made acceot/rehect functions, which change the accepted boolean on the bid.