Solved the check of accepted orders by workers
- Added 2 schemas : checking and note
- Added a route: checkings
- Added the route in the server.js

Now we can make a checking, based on an existing order. On that checking we can create notes. Each note is marked with its date of creation.
If the workers finished their job, they can mark the checking with a boolean: finished.
There are 2 options to see an order's details.
We can check the whole order, or we can check only the orders items(garage doors).